[issue/69] - preventing duplicated data from client
in the database, since the client can (yet) send the same
data multiple times.

Because of this in the JSON there are the uuid fields which have the same
values for the same entrys. Some of the database models had to be
changed and a UUID field had to be added there to hold the uuid to later
exclude data recieved with the same uuid.